+++ modules/server/src/index.js
@@ -16,6 +16,7 @@ import {
   addMappingsToTypes,
   mappingToAggsState,
 } from '@arranger/mapping-utils';
+import projectsRoutes from './projects';
 
 let writeFile = promisify(fs.writeFile);
 
@@ -52,10 +53,16 @@ let main = async () => {
   let http = Server(app);
   let io = socketIO(http);
 
-  let connections = [];
+  let connections = {};
+  let projectStates = [];
 
   io.on('connection', socket => {
-    connections.push(socket);
+    connections[socket.id] === socket;
+
+    socket.on('disconnect', () => {
+      delete connections[socket.id];
+      clearInterval(socket.monitorIntervalId);
+    });
 
     socket.on('arranger::monitorProjects', ({ projects = [], eshost }) => {
       socket.monitorIntervalId = setInterval(async () => {
@@ -103,382 +110,7 @@ let main = async () => {
     next();
   });
 
-  app.use('/projects/:id/types/:index/fields', async (req, res) => {
-    let { es } = req.context;
-    let { id, index } = req.params;
-
-    if (!id || !index) {
-      return res.json({ error: 'missing fields' });
-    }
-
-    let arrangerconfig = {
-      projectsIndex: {
-        index: `arranger-projects-${id}-${index}`,
-        type: `arranger-projects-${id}-${index}`,
-      },
-    };
-
-    let fields = [];
-
-    try {
-      fields = await es.search(arrangerconfig.projectsIndex);
-    } catch (error) {
-      try {
-        await es.indices.create({
-          index: arrangerconfig.projectsIndex.index,
-        });
-
-        // TODO: check for aliases!!
-        let mappings = await es.indices.getMapping({
-          index,
-          type: index,
-        });
-
-        let mapping = mappings[index].mappings[index].properties;
-
-        let fields = mappingToAggsState(mapping);
-
-        let body = flattenDeep(
-          fields.map(x => [
-            {
-              index: {
-                _index: arrangerconfig.projectsIndex.index,
-                _type: arrangerconfig.projectsIndex.index,
-                _id: x.field,
-              },
-            },
-            JSON.stringify(x),
-          ]),
-        );
-
-        await es.bulk({ body });
-        return res.json({ fields, total: fields.length });
-      } catch (error) {
-        return res.json({ error: error.message });
-      }
-      return res.json({ error: error.message });
-    }
-
-    res.json({ fields: mapHits(fields), total: fields.hits.total });
-  });
-
-  app.use('/projects/:id/types/add', async (req, res) => {
-    let { es } = req.context;
-    let { id } = req.params;
-    let { index, name } = req.body;
-
-    if (!id || !index || !name) {
-      return res.json({ error: 'missing fields' });
-    }
-
-    let arrangerconfig = {
-      projectsIndex: {
-        index: `arranger-projects-${id}`,
-        type: `arranger-projects-${id}`,
-      },
-    };
-
-    try {
-      await es.create({
-        ...arrangerconfig.projectsIndex,
-        refresh: true,
-        id: index,
-        body: {
-          index,
-          name,
-          active: true,
-          timestamp: new Date().toISOString(),
-        },
-      });
-    } catch (error) {
-      return res.json({ error: error.message });
-    }
-
-    let types = [];
-
-    try {
-      types = await es.search(arrangerconfig.projectsIndex);
-    } catch (error) {
-      try {
-        await es.indices.create({
-          index: arrangerconfig.projectsIndex.index,
-        });
-        return res.json({ types });
-      } catch (error) {
-        return res.json({ error: error.message });
-      }
-      return res.json({ error: error.message });
-    }
-
-    let hits = mapHits(types);
-
-    let mappings = await fetchMappings({ es, types: hits });
-
-    res.json({
-      types: hits.map(x => ({
-        ...x,
-        mappings: mappings.find(y => y.index === x.index).mapping,
-      })),
-      total: types.hits.total,
-    });
-  });
-
-  app.use('/projects/:id/spinup', async (req, res) => {
-    let { es } = req.context;
-    let { id } = req.params;
-    if (!id) return res.json({ error: 'project empty' });
-
-    let arrangerconfig = {
-      projectsIndex: {
-        index: `arranger-projects`,
-        type: `arranger-projects`,
-        id,
-        body: {
-          doc: {
-            active: true,
-          },
-        },
-      },
-    };
-
-    try {
-      await es.update(arrangerconfig.projectsIndex);
-    } catch (error) {
-      return res.json({ error: error.message });
-    }
-
-    arrangerconfig = {
-      projectsIndex: {
-        index: `arranger-projects-${id}`,
-        type: `arranger-projects-${id}`,
-      },
-    };
-
-    let types = [];
-
-    try {
-      types = await es.search(arrangerconfig.projectsIndex);
-    } catch (error) {
-      try {
-        await es.indices.create({
-          index: arrangerconfig.projectsIndex.index,
-        });
-        return res.json({ types });
-      } catch (error) {
-        return res.json({ error: error.message });
-      }
-      return res.json({ error: error.message });
-    }
-
-    let hits = mapHits(types);
-    let mappings = await fetchMappings({ es, types: hits });
-
-    let typesWithMappings = addMappingsToTypes({
-      types: hits.map(type => [
-        type.index,
-        {
-          index: type.index,
-          es_type: type.index,
-          name: type.name,
-          customFields: ``,
-        },
-      ]),
-      mappings: mappings.map(m => m.mapping),
-    });
-
-    let schema = makeSchema({ types: typesWithMappings, rootTypes: [] });
-
-    app.get(`/${id}/ping`, (req, res) => res.send('ok'));
-
-    app.use(
-      `/${id}/graphql`,
-      schema
-        ? graphqlExpress({ schema, context: { es } })
-        : (req, res) =>
-            res.json({
-              error:
-                'schema is undefined. Make sure you provide a valid GraphQL Schema. https://www.apollographql.com/docs/graphql-tools/generate-schema.html',
-            }),
-    );
-
-    console.log(`graphql server running at ${id}/graphql`);
-
-    res.json({ message: `graphql server running at ${id}/graphql` });
-  });
-
-  app.use('/projects/:id/types', async (req, res) => {
-    let { es } = req.context;
-    let { id } = req.params;
-    if (!id) return res.json({ error: 'project empty' });
-
-    let arrangerconfig = {
-      projectsIndex: {
-        index: `arranger-projects-${id}`,
-        type: `arranger-projects-${id}`,
-      },
-    };
-
-    let types = [];
-
-    try {
-      types = await es.search(arrangerconfig.projectsIndex);
-    } catch (error) {
-      try {
-        await es.indices.create({
-          index: arrangerconfig.projectsIndex.index,
-        });
-        return res.json({ types });
-      } catch (error) {
-        return res.json({ error: error.message });
-      }
-      return res.json({ error: error.message });
-    }
-
-    let hits = mapHits(types);
-
-    let mappings = await fetchMappings({ es, types: hits });
-
-    res.json({
-      types: hits.map(x => ({
-        ...x,
-        mappings: mappings.find(y => y.index === x.index).mapping,
-      })),
-      total: types.hits.total,
-    });
-  });
-
-  app.use('/projects/:id/delete', async (req, res) => {
-    let { es } = req.context;
-    let { id } = req.params;
-
-    if (!id) return res.json({ error: 'id cannot be empty' });
-
-    let projects = [];
-
-    let arrangerconfig = {
-      projectsIndex: {
-        index: 'arranger-projects',
-        type: 'arranger-projects',
-      },
-    };
-
-    try {
-      await es.delete({
-        ...arrangerconfig.projectsIndex,
-        refresh: true,
-        id,
-      });
-      await es.indices.delete({
-        index: `arranger-projects-${id}`,
-      });
-    } catch (error) {
-      return res.json({ error: error.message });
-    }
-
-    try {
-      projects = await es.search({
-        ...arrangerconfig.projectsIndex,
-        size: 1000,
-      });
-    } catch (error) {
-      try {
-        await es.indices.create({
-          index: arrangerconfig.projectsIndex.index,
-        });
-        return res.json({ projects });
-      } catch (error) {
-        return res.json({ error: error.message });
-      }
-      return res.json({ error: error.message });
-    }
-
-    res.json({ projects: mapHits(projects), total: projects.hits.total });
-  });
-
-  app.use('/projects/add', async (req, res) => {
-    let { es } = req.context;
-    let { id } = req.body;
-
-    if (!id) return res.json({ error: 'id cannot be empty' });
-
-    let projects = [];
-
-    let arrangerconfig = {
-      projectsIndex: {
-        index: 'arranger-projects',
-        type: 'arranger-projects',
-      },
-    };
-
-    try {
-      await es.create({
-        ...arrangerconfig.projectsIndex,
-        refresh: true,
-        id,
-        body: {
-          id,
-          active: false,
-          timestamp: new Date().toISOString(),
-        },
-      });
-    } catch (error) {
-      return res.json({ error: error.message });
-    }
-
-    try {
-      projects = await es.search({
-        ...arrangerconfig.projectsIndex,
-        size: 1000,
-      });
-    } catch (error) {
-      try {
-        await es.indices.create({
-          index: arrangerconfig.projectsIndex.index,
-        });
-        return res.json({ projects });
-      } catch (error) {
-        return res.json({ error: error.message });
-      }
-      return res.json({ error: error.message });
-    }
-
-    res.json({ projects: mapHits(projects), total: projects.hits.total });
-  });
-
-  app.use('/projects', async (req, res) => {
-    let { es } = req.context;
-    // create arranger-projects index
-    // get arranger projects
-
-    let projects = [];
-
-    let arrangerconfig = {
-      projectsIndex: {
-        index: 'arranger-projects',
-        type: 'arranger-projects',
-        size: 1000,
-      },
-    };
-
-    try {
-      projects = await es.search(arrangerconfig.projectsIndex);
-    } catch (error) {
-      try {
-        await es.indices.create({
-          index: arrangerconfig.projectsIndex.index,
-        });
-        res.json({ projects });
-      } catch (error) {
-        return res.json({ error: error.message });
-      }
-      return res.json({ error: error.message });
-    }
-
-    res.json({
-      projects: projects.hits.hits.map(x => x._source),
-      total: projects.hits.total,
-    });
-  });
+  projectsRoutes({ app });
 
   http.listen(port, () => rainbow(`⚡️ Listening on port ${port} ⚡️`));
 };
